Added support for UITableViewCell
UITableViewCell calls .backgroundColor = [UIColor whiteColor] on all
subviews when selected. This removes the red circle if the notification
hub is located in a tableviewcell.

I created a small subclass of UIView where I overrode
setBackgroundColor and only allows it to be set if the user(programmer)
explicitly has given his permission. The system cannot set the
backgroundcolor because it is not aware of this ivar.